Rewrite mowgli_patricia to reduce cache misses.
Keys are now analyzed per group of 4 bits instead of
per bit, so each node access (possible cache miss)
processes up to 4 bits of key instead of just one.
Because the differing bits may not be adjacent, the
branching factor may be less than 16, but it will be
at least 2. The worst case for the depth of the tree
is one fourth of that of the old code.

As a side effect, the upward links of the old version
are no longer possible and have been removed. There
are now distinct nodes and leaves, and there may be
many more leaves than nodes. This should make the code
easier to understand.

A parent pointer in each node and leaf has replaced the
linked list of all elements. This provides similar
functionality with one less pointer and better
integration with the tree.

As a side effect, the iterators now return the elements
in order.

The new code is believed to be ABI and API compatible
with the old code. In fact, mowgli_patricia.h has not
changed at all.