base/memslice: new allocator.
This is a magazine-allocator using a global set of private heaps, similar to
sharedheap (which is implemented in a few mowgli-using applications).

It has a few advantages over sharedheap, though, such as O(1) operations and
the lack of requirement for refcounting the heaps.  As it implements
mowgli.core.allocation_policy, it can be used as a mowgli_alloc() implementation.