linebuf: Add shutdown semantics
A linebuf that has been shut down will continue to perform normally,
but will stop calling the readline callback, and will call a shutdown
callback when all data is finished being written.

The addition to the linebuf struct means that applications using linebuf
will need to be re-compiled to the new definition.