Fix IPv6 reverse DNS lookups (address to name conversion)
The bug here is that it reverses the octets and appends a period
after each nibble, but then prepends a period before ip6.arpa.

This results in a query with 2 consecutive periods, which then
goes on to fail query validation and so the query is never sent.

I have no idea how this failed basic quality testing especially
since there is an example program that uses this functionality
that was apparently not even extended to try IPv6.