All Header Files: Use consistently-named & commented inclusion guards
This names the include guards after the file path and comments the
endif at the bottom of the files too. It also instructs autoconf to
generate inclusion guards for the autoconf.h file that it writes out.

This avoids the include guards being named with reserved identifiers.
Identifiers that start with 2 underscores, or 1 underscore and then a
capital letter, should only be declared by system headers.