container/list: mowgli_list_concat(): fix possible NULL dereference
When concatenating a list onto an empty list, the code would dereference
the empty list's "tail" member even after checking that it is not NULL
(but not within this check, hence the problem). Move the dereference to
inside the conditional check block.

Also ensure that if the list being concatenated onto is empty, that its
"head" member is set appropriately too.

The former issue was identified by the Clang Static Analyzer.

Finally remove an "a = b = ..." idiom that trashes readability.