patricia: add a mowgli_patricia_clear() function
This does the same as mowgli_patricia_destroy(), but without actually
destroying the trie itself. Useful when you need to clear all of the
elements in a trie without losing the trie too.

You cannot always use MOWGLI_PATRICIA_FOREACH to do this, as it
returns the items on each iteration, not their key. If the item
doesn't contain the key, this is not useful.

Redefine mowgli_patricia_destroy() in terms of the new function.